+++ src/app/app.module.ts
@@ -12,10 +12,18 @@ import { RasTokenCallbackModule } from './modules/ras-token-callback/ras-token-c
 import { LaunchModule } from './modules/launch/launch.module';
 import { InitializeSpinnerModule } from './shared/initialize-spinner/initialize-spinner.module';
 import { AnnouncerModule } from './shared/announcer/announcer.module';
+import {
+  MAT_TOOLTIP_DEFAULT_OPTIONS,
+  MAT_TOOLTIP_DEFAULT_OPTIONS_FACTORY
+} from "@angular/material/tooltip";
+import {Oauth2TokenCallbackModule} from "./modules/oauth2-token-callback/oauth2-token-callback.module";
 
 const appearance: MatFormFieldDefaultOptions = {
   appearance: 'outline'
 };
+
+const myCustomTooltipDefaults = MAT_TOOLTIP_DEFAULT_OPTIONS_FACTORY();
+myCustomTooltipDefaults.disableTooltipInteractivity = true;
 
 /**
  * Application initializer.
@@ -33,6 +41,7 @@ function initializeApp(settingsService: SettingsService): () => Promise<any> {
     AppRoutingModule,
     HomeModule,
     RasTokenCallbackModule,
+    Oauth2TokenCallbackModule,
     LaunchModule,
     InitializeSpinnerModule
   ],
@@ -40,6 +49,10 @@ function initializeApp(settingsService: SettingsService): () => Promise<any> {
     {
       provide: MAT_FORM_FIELD_DEFAULT_OPTIONS,
       useValue: appearance
+    },
+    {
+      provide: MAT_TOOLTIP_DEFAULT_OPTIONS,
+      useValue: myCustomTooltipDefaults
     },
     {
       provide: APP_INITIALIZER,