+++ src/app/shared/fhir-backend/fhir-batch-query.js
@@ -8,6 +8,8 @@ const HTTP_ABORT = 0;
 const UNSUPPORTED_VERSION = -1;
 // The value of property status in the rejection object when the "metadata" query indicated basic authentication failure
 const BASIC_AUTH_REQUIRED = -2;
+// The value of property status in the rejection object when the "metadata" query indicated that OAuth2 is required
+const OAUTH2_REQUIRED = -3;
 // Request priorities (numbers by which requests in the pending queue are sorted)
 const PRIORITIES = {
   LOW: 100,
@@ -186,7 +188,7 @@ class FhirBatchQuery extends EventTarget {
     // const currentServiceBaseUrl = this._serviceBaseUrl;
 
     if (!this._initializationPromise) {
-      this._features = {};
+      this._features = {isFormatSupported: true};
       this._initializationPromise = this.makeInitializationCalls(predefinedInitResult);
     }
     return this._initializationPromise;
@@ -207,39 +209,8 @@ class FhirBatchQuery extends EventTarget {
     // Common options for initialization requests
     // retryCount=2, We should not try to resend the first request to the server many times - this could be the wrong URL
     const options = this.getCommonInitRequestOptions();
-    // useInitContext=false, The request is cached as the same name before and after login, so we don't make the request again after login.
-    const options_noInitContext = this.getCommonInitRequestOptions(false);
 
-    // Below are initialization requests that are always made.
-    const initializationRequests =
-      // Retrieve the information about a server's capabilities (https://www.hl7.org/fhir/http.html#capabilities)
-      // Do not cache /metadata query that requires basic authentication.
-      // Credentials must be re-entered if user closes and opens the browser.
-      (this.initContext === 'basic-auth'
-        ? this.get('metadata?_elements=fhirVersion', options)
-        : this.getWithCache('metadata?_elements=fhirVersion', options_noInitContext))
-      .then((metadata) => {
-        const fhirVersion = metadata.data.fhirVersion;
-        this._versionName = getVersionNameByNumber(fhirVersion);
-        if (!this._versionName) {
-          return Promise.reject({
-            status: UNSUPPORTED_VERSION, error: 'Unsupported FHIR version: ' + fhirVersion
-          });
-        }
-      }, (metadata) => {
-        // If initialization fails, do not cache initialization responses
-        this.clearCacheByName(this.getInitCacheName());
-        // Abort other initialization requests
-        this.clearPendingRequests();
-        if (metadata.status === 401 && metadata.wwwAuthenticate?.startsWith('Basic')) {
-          return Promise.reject({
-            status: BASIC_AUTH_REQUIRED, error: 'basic authorization required'
-          });
-        }
-        return Promise.reject({
-          error: 'Could not retrieve the FHIR server\'s metadata. Please make sure you are entering the base URL for a FHIR server.'
-        });
-      })
+    const initializationRequests = this.checkMetadata()
       .then(() => {
         if (predefinedInitResult && predefinedInitResult.version === this._versionName) {
           this._features = predefinedInitResult.features;
@@ -256,7 +227,16 @@ class FhirBatchQuery extends EventTarget {
             url: this._serviceBaseUrl,
             body: JSON.stringify({
               resourceType: 'Bundle',
-              type: 'batch'
+              type: 'batch',
+              // Include an entry of the metadata query that we know would succeed if sent separately.
+              // The server might return status === "400 Bad Request" for the entry, even though the POST request
+              // itself returns with a 200, in which case the server does not support batch request.
+              entry: [{
+                request: {
+                  method: 'GET',
+                  url: this.metaDataQuery
+                }
+              }]
             }),
             logPrefix: 'Batch',
             combine: false,
@@ -274,10 +254,14 @@ class FhirBatchQuery extends EventTarget {
             hasResearchStudy.status === 'fulfilled' &&
             hasResearchStudy.value.data.entry?.length > 0;
           this._features.missingModifier = missingModifier.status === 'fulfilled';
-          this._features.batch = batch.status === 'fulfilled';
+          this._features.batch =
+            batch.status === 'fulfilled' &&
+            batch.value.data.entry &&
+            batch.value.data.entry.length > 0 &&
+            batch.value.data.entry[0].response?.status?.startsWith('200');
         })
         .then(() => {
-          // On dbGaP server, only do initializationRequests2 requests after login.
+          // On dbGaP server, only do certain initialization requests after login.
           if (this.initContext === 'dbgap-pre-login') {
             // Check if server has at least one Research Study with Research Subjects.
             // No need to make this request if there is no Research Study at all.
@@ -335,6 +319,63 @@ class FhirBatchQuery extends EventTarget {
       });
 
     return initializationRequests;
+  }
+
+  /**
+   * Retrieve the information about a server's capabilities (https://www.hl7.org/fhir/http.html#capabilities)
+   * @param {boolean} withElementsParam - whether to include _elements parameter in the metadata query.
+   * @returns {Promise<void>}
+   */
+  checkMetadata(withElementsParam = true) {
+    // retryCount=2, We should not try to resend the first request to the server many times - this could be the wrong URL
+    const options = this.getCommonInitRequestOptions();
+    // useInitContext=false, The request is cached as the same name before and after login, so we don't make the request again after login.
+    const options_noInitContext = this.getCommonInitRequestOptions(false);
+    const elementsParam = withElementsParam ? '?_elements=fhirVersion' : '';
+    // Do not cache /metadata query that requires basic authentication.
+    // Credentials must be re-entered if user closes and opens the browser.
+    return (this.initContext === 'basic-auth'
+      ? this.get(`metadata${elementsParam}`, options)
+      : this.getWithCache(`metadata${elementsParam}`, options_noInitContext))
+    .then((metadata) => {
+      const fhirVersion = metadata.data.fhirVersion;
+      this._versionName = getVersionNameByNumber(fhirVersion);
+      if (!this._versionName) {
+        return Promise.reject({
+          status: UNSUPPORTED_VERSION, error: 'Unsupported FHIR version: ' + fhirVersion
+        });
+      }
+      // Mark the metadata query url for later use in the batch query.
+      this.metaDataQuery = `metadata${elementsParam}`;
+      if (this._features.isFormatSupported) {
+        this.metaDataQuery += withElementsParam ? '&_format=json' : '?_format=json';
+      }
+    }, (metadata) => {
+      // If initialization fails, do not cache initialization responses
+      this.clearCacheByName(this.getInitCacheName());
+      // Abort other initialization requests
+      this.clearPendingRequests();
+      if (metadata.status === 401 && metadata.wwwAuthenticate?.startsWith('Basic')) {
+        return Promise.reject({
+          status: BASIC_AUTH_REQUIRED, error: 'basic authentication required'
+        });
+      }
+      if (metadata.status === 401 && metadata.wwwAuthenticate?.startsWith('Bearer')) {
+        return Promise.reject({
+          status: OAUTH2_REQUIRED, error: 'oauth2 required'
+        });
+      }
+      this._features.isFormatSupported = !metadata.error.includes('_format');
+      const errorIncludesElements = metadata.error.includes('_elements');
+      if (errorIncludesElements || !this._features.isFormatSupported) {
+        // If it complains about '_format', set _features.isFormatSupported and make the /metadata request again.
+        // If it complains about '_elements', make the /metadata request again without the '_elements' parameter.
+        return this.checkMetadata(!errorIncludesElements);
+      }
+      return Promise.reject({
+        error: 'Could not retrieve the FHIR server\'s metadata. Please make sure you are entering the base URL for a FHIR server.'
+      });
+    })
   }
 
   /**
@@ -735,7 +776,7 @@ class FhirBatchQuery extends EventTarget {
         sendUrl.searchParams.append('api_key', this._apiKey);
       }
 
-      if (!sendUrl.searchParams.has('_format')) {
+      if (this._features.isFormatSupported && !sendUrl.searchParams.has('_format')) {
         sendUrl.searchParams.append('_format', 'json');
       }
 
@@ -917,7 +958,7 @@ class FhirBatchQuery extends EventTarget {
       return data.issue.map((item) => item.diagnostics).join('\n') || '';
     }
 
-    return 'Unknown Error';
+    return data?.error?.message || 'Unknown Error';
   }
 
   clearPendingRequests() {